+++ launcher/minecraft/auth/AccountList.h
@@ -44,7 +44,7 @@
 
 /*!
  * List of available Mojang accounts.
- * This should be loaded in the background by PolyMC on startup.
+ * This should be loaded in the background by Prism Launcher on startup.
  */
 class AccountList : public QAbstractListModel
 {
