+++ launcher/minecraft/auth/MinecraftAccount.h
@@ -61,7 +61,7 @@ Q_DECLARE_METATYPE(MinecraftAccountPtr)
  * A profile within someone's Mojang account.
  *
  * Currently, the profile system has not been implemented by Mojang yet,
- * but we might as well add some things for it in PolyMC right now so
+ * but we might as well add some things for it in Prism Launcher right now so
  * we don't have to rip the code to pieces to add it later.
  */
 struct AccountProfile
