+++ launcher/minecraft/launch/MinecraftServerTarget.cpp
@@ -23,7 +23,7 @@ MinecraftServerTarget MinecraftServerTarget::parse(const QString &fullAddress) {
 
     // The logic below replicates the exact logic minecraft uses for parsing server addresses.
     // While the conversion is not lossless and eats errors, it ensures the same behavior
-    // within Minecraft and PolyMC when entering server addresses.
+    // within Minecraft and Prism Launcher when entering server addresses.
     if (fullAddress.startsWith("["))
     {
         int bracket = fullAddress.indexOf("]");
