+++ launcher/settings/Setting.h
@@ -33,7 +33,7 @@ public:
      * Construct a Setting
      *
      * Synonyms are all the possible names used in the settings object, in order of preference.
-     * First synonym is the ID, which identifies the setting in PolyMC.
+     * First synonym is the ID, which identifies the setting in Prism Launcher.
      *
      * defVal is the default value that will be returned when the settings object
      * doesn't have any value for this setting.
