+++ launcher/ui/MainWindow.cpp
@@ -1860,7 +1860,7 @@ void MainWindow::globalSettingsClosed()
     updateToolsMenu();
     updateStatusCenter();
     // This needs to be done to prevent UI elements disappearing in the event the config is changed
-    // but PolyMC exits abnormally, causing the window state to never be saved:
+    // but Prism Launcher exits abnormally, causing the window state to never be saved:
     APPLICATION->settings()->set("MainWindowState", saveState().toBase64());
     update();
 }
