+++ libraries/launcher/org/prismlauncher/LauncherFactory.java
@@ -33,10 +33,10 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.polymc;
+package org.prismlauncher;
 
-import org.polymc.impl.OneSixLauncher;
-import org.polymc.utils.Parameters;
+import org.prismlauncher.impl.OneSixLauncher;
+import org.prismlauncher.utils.Parameters;
 
 import java.util.HashMap;
 import java.util.Map;
