+++ libraries/launcher/org/prismlauncher/utils/Parameters.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package org.polymc.utils;
+package org.prismlauncher.utils;
 
-import org.polymc.exception.ParameterNotFoundException;
+import org.prismlauncher.exception.ParameterNotFoundException;
 
 import java.util.ArrayList;
 import java.util.HashMap;
